Fix yet another warning.
git-svn-id: http://svn.inspircd.org/repository/trunk/inspircd@9271 e03df62e-2008-0410-955e-edbf42e46eb7

+++ src/inspircd.cpp
@@ -651,7 +651,7 @@ int InspIRCd::Run()
 		{
 			if (TIME < OLDTIME)
 			{
-				SNO->WriteToSnoMask('A', "\002EH?!\002 -- Time is flowing BACKWARDS in this dimension! Clock drifted backwards %ld secs.",OLDTIME-TIME);
+				SNO->WriteToSnoMask('A', "\002EH?!\002 -- Time is flowing BACKWARDS in this dimension! Clock drifted backwards %lu secs.", (unsigned long)OLDTIME-TIME);
 			}
 
 			if ((TIME % 3600) == 0)
